+++ src/App.scss
@@ -145,12 +145,12 @@ header {
   }
 }
 
-@media only screen and (min-width: 340px) and (max-width: 768px) {
+@media only screen and (min-width: 320px) and (max-width: 768px) {
   .anime-list {
     width: 100%;
     padding: 1rem 4rem;
     display: grid;
-    grid-template-columns: repeat(2, 1fr);
+    grid-template-columns: repeat(3, 1fr);
     grid-gap: 10px;
 
     .suggestions {
